Use the newer vue syntax for deep CSS selector
I don't know why this is necessary (there's talk of different syntax in vue3, but this should still work in vue2), but it is - without this, the resulting rules are absent from `dist/inbox-component.esm.js`. With this change, the rules get correctly compiled as e.g. `[data-v-511a5242] .cursor-pointer{cursor:pointer}`.