Use a fixed size for images in the message/inbox view
If images are a variable size, weird things happen to scroll position as images load asynchronously. With this approach, the image is always the same size (and square), and it hides parts of the image that are outside the center square. If you click into the modal or view it in gallery mode, it displays the full image.